requirements.txt
torch
torchvision
torchaudio
transformers
Pillow
faiss-cpu
flask
boto3
